requirements.txt
hikari~=2.0.0.dev106
hikari-tanjun~=2.4.0a1
hikari-yuyo>=1.0.5a1.post1

aioredis~=2.0.0
aiohttp~=3.8.1
